Made the Spotify title tracker
Did it all in one fell swoop across 2 nights. Forgot to commit haha. Basically tracks the window title of spotify. Recovers when spotifies closes/opens & music is paused. Does error logging & has a customizable config in the documents folder that is generated. Uses the spotify API to get the real name of the track + artists & album art. otherwise it defaults to the window title. Need to add gui form for the config as well.